Changed UART behaviour to be standard by default
Using the standard UART options and device tree overlays is probably a
sensible default instead of changing UART options in our own overlay.
The 'uart_disable' parameter is now 'bluetooth_disable', which will
enable the main Raspberry Pi UART for serial and disable Bluetooth (the
old default in this overlay).