Added ability to power off board automatically when using aux battery
This is achieved through a combination of config.txt options and systemd
shutdown script. The config.txt GPIO options sets the input mode of the power
control pin to pull-down, then immediately sets the pin to output high. This
keeps the board powered on, but means the pin goes low when the CPU is halted
(so after shutdown). This would work on its own, but also means the board turns
off after a 'reboot' or 'halt' command (which it ideally shouldn't). The
systemd shutdown script will set the power control pin to pull-up instead if
either of these shutdowns happens, and the board will not power off. The script
also reinforces the pull-down on 'poweroff', in case it has been changed during
runtime.

As the gpio-poweroff kernel driver is not necessary for driving this pin at
poweroff, it has also been removed completely from the device tree.