Updated utilities and examples to support latest Debian Buster kernel
Linux kernel 5.4 (or possibly an earlier release) changed the numbering
of I2C switch downstream buses from 3-10 to 11-18, and replaced the
hwmon driver for the ADS1015 ADC with an iio driver.

This commit updates all examples and utilities in the repository to use
the new bus numbers and ADC driver by default, but each will
automatically fall back to the old systems if necessary.